pyproject.toml
[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "online-cinema"
version = "0.1.0"
description = "Online cinema API (FastAPI)"
authors = ["Online Cinema Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.115"
uvicorn = { version = ">=0.30", extras = ["standard"] }
email-validator = ">=2.0"
pydantic-settings = ">=2.0"
sqlalchemy = ">=2.0"
aiosqlite = ">=0.19"
greenlet = ">=3.0"
python-jose = { version = ">=3.3", extras = ["cryptography"] }
passlib = { version = ">=1.7", extras = ["bcrypt"] }
bcrypt = ">=4.0.1,<4.1"
python-multipart = ">=0.0.9"
stripe = ">=11.0,<12"
celery = { version = ">=5.3,<6", extras = ["redis"] }
redis = ">=5,<6"

[tool.poetry.group.dev.dependencies]
pytest = ">=8,<9"
pytest-asyncio = ">=0.24,<1"
httpx = ">=0.27,<1"
black = ">=24,<25"
isort = ">=5.13,<6"
flake8 = ">=7,<8"
flake8-bugbear = ">=24,<25"
mypy = ">=1.11,<2"
types-passlib = "*"
types-python-jose = "*"
pytest-cov = "^7.1.0"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src"]
known_first_party = ["src"]

[tool.mypy]
python_version = "3.12"
files = ["src"]
explicit_package_bases = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "src.tests.*"
ignore_errors = true

[tool.coverage.run]
omit = ["src/tests/*"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

# macOS: if `poetry install` fails with [Errno 2] No such file or directory: 'python',
# use: ./scripts/poetry.sh install --with dev   (or: deactivate && rm -rf .venv && retry)
